Write steam helper function to dry code
